+++ README.md
@@ -14,17 +14,15 @@ You can start editing the page by modifying `app/page.tsx`. The page auto-update
 
 This project uses [`next/font`](https://nextjs.org/docs/app/building-your-application/optimizing/fonts) to automatically optimize and load [Geist](https://vercel.com/font), a new font family for Vercel.
 
-## Learn More
+## Components
 
-To learn more about Next.js, take a look at the following resources:
+- Resume viewer, parser, and html generator.
+- MongoDB to manage text content; quotes and bio.
+- The Wilting Demo, a unity game I am making. _STILL UNDER CONSTRUCTION ETA 3/24/25_
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+## Secrets
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js) - your feedback and contributions are welcome!
-
-## Deploy on Vercel
-
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/app/building-your-application/deploying) for more details.
+- NEXT_PUBLIC_CURRENT_RESUME: Location of the current Resume PDF.
+- NEXT_PUBLIC_UNITY_ENABLED: Feature flag to enable Unity WebGL player.
+- MONGODB_URI: MongoDB connection string.
+- MONGODB_DB: Collection name.
